Improve Project Structure and Analytics Service
