+++ __tests__/setup.ts
@@ -1,1 +1,7 @@
 import '@testing-library/jest-dom';
+import { afterEach } from 'vitest';
+import { analytics } from '../src/services/analytics';
+
+afterEach(() => {
+	analytics.reset();
+});
