+++ src/components/partials/Cyrillic.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import Cyrillic from './Cyrillic';
-import TransliterateContextProvider from '../contexts/TransliterateContext';
+import TransliterateContextProvider from '../../contexts/TransliterateContext';
 
 describe('Cyrillic component', () => {
 	it('renders textarea with correct attributes', () => {
