+++ src/components/partials/Latin.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import Latin from './Latin';
-import TransliterateContextProvider from '../contexts/TransliterateContext';
+import TransliterateContextProvider from '../../contexts/TransliterateContext';
 
 describe('Latin component', () => {
 	it('renders textarea with correct attributes', () => {
