+++ src/contexts/TransliterateContext.tsx
@@ -8,8 +8,8 @@ import {
 	RefObject
 } from 'react';
 import transliterate from 'serbian-transliterate';
-import containsUpperCase from '../../helpers/containsUpperCase';
-import { trackTransliteration } from '../../utils/analytics';
+import containsUpperCase from '../helpers/containsUpperCase';
+import { analytics } from '../services/analytics';
 
 interface ITransliterateContext {
 	cyrillic: string;
@@ -73,18 +73,21 @@ function TransliterateContextProvider({ children }: { children: ReactNode }) {
 		} else if (lastEdit === 'latin') {
 			setCyrillic(transliterate(latin, 'toCyrillic'));
 		}
+	}, [cyrillic, latin, lastEdit]);
 
-		// Only set up tracking if there's a valid lastEdit
-		if (lastEdit === 'cyrillic' || lastEdit === 'latin') {
-			// Track transliteration usage with debouncing (only track after 2 seconds of inactivity)
-			if (trackingTimeoutRef.current !== null) {
-				clearTimeout(trackingTimeoutRef.current);
-			}
+	useEffect(() => {
+		// Debounce tracking to avoid firing on every keystroke
+		if (trackingTimeoutRef.current !== null) {
+			clearTimeout(trackingTimeoutRef.current);
+		}
 
+		// Only track if there was an edit and there is text
+		if ((lastEdit === 'cyrillic' || lastEdit === 'latin')) {
 			const text = lastEdit === 'cyrillic' ? cyrillic : latin;
+
 			if (text.length > 0) {
 				trackingTimeoutRef.current = window.setTimeout(() => {
-					trackTransliteration({
+					analytics.trackTransliteration({
 						direction: lastEdit === 'cyrillic' ? 'toLatin' : 'toCyrillic',
 						textLength: text.length
 					});
@@ -92,7 +95,6 @@ function TransliterateContextProvider({ children }: { children: ReactNode }) {
 			}
 		}
 
-		// Always return cleanup function
 		return () => {
 			if (trackingTimeoutRef.current !== null) {
 				clearTimeout(trackingTimeoutRef.current);
